Changed the way we store captions on the initial history retrieving
